SFX, Audio Scripts, and more.
This commit includes small modifications to add audio functionality to
the Game Manager script, Flower Squisher script, Dialogue System script
and I’ve updated the Action and Movement Audio script. I’ve also added
a random audio sample trigger script which is used on the ice platform,
flowers, and some others. I’ve added play on awake sounds to some
prefabs like the drinks and the snowballs. I added some reverb zones
and a few blank cubes to act as triggers or to provide different
surfaces like in Lakes for the snow on the left. Not a permanent
solution. I also added a script that isn’t implemented yet (will do
this week once I put in the Audio Mixers and route everything) that
puts a filter over the sound effects and music while crafting.